backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aioredis==2.0.1
weaviate-client==3.25.3
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6
